Xsv.Test sample files made embedded.   - VSO PR build unit tests aren't finding "copy to output" files; changing to use embedded resources.
